extrace check_compacted to test_raftstore
