+++ tests/integrations/raftstore/test_compact_log.rs
@@ -44,46 +44,6 @@ fn test_compact_log<T: Simulator>(cluster: &mut Cluster<T>) {
     panic!("after inserting 1000 entries, compaction is still not finished.");
 }
 
-fn check_compacted(
-    all_engines: &HashMap<u64, Engines<RocksEngine, RocksEngine>>,
-    before_states: &HashMap<u64, RaftTruncatedState>,
-    compact_count: u64,
-) -> bool {
-    // Every peer must have compacted logs, so the truncate log state index/term must > than before.
-    let mut compacted_idx = HashMap::default();
-
-    for (&id, engines) in all_engines {
-        let mut state: RaftApplyState = get_raft_msg_or_default(engines, &keys::apply_state_key(1));
-        let after_state = state.take_truncated_state();
-
-        let before_state = &before_states[&id];
-        let idx = after_state.get_index();
-        let term = after_state.get_term();
-        if idx == before_state.get_index() || term == before_state.get_term() {
-            return false;
-        }
-        if idx - before_state.get_index() < compact_count {
-            return false;
-        }
-        assert!(term > before_state.get_term());
-        compacted_idx.insert(id, idx);
-    }
-
-    // wait for actual deletion.
-    sleep_ms(100);
-
-    for (id, engines) in all_engines {
-        for i in 0..compacted_idx[id] {
-            let key = keys::raft_log_key(1, i);
-            if engines.raft.get_value(&key).unwrap().is_none() {
-                break;
-            }
-            assert!(engines.raft.get_value(&key).unwrap().is_none());
-        }
-    }
-    true
-}
-
 fn test_compact_count_limit<T: Simulator>(cluster: &mut Cluster<T>) {
     cluster.cfg.raft_store.raft_log_gc_count_limit = 100;
     cluster.cfg.raft_store.raft_log_gc_threshold = 500;
